+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Version 0.24
 - `dl_directory` check constraint that dirpath is like '/%/'
 - new `is_local()`, `is_valid()` as `is_http_success()` immutable functions, for use in check constraints and such
 - forbid * in file names
+- new `iri()` function to convert IRIs (unicode) to URIs (escaped) 
+- `dlvalue()` now supports `IRI` link type, bug fixes
 
 Version 0.23
 ------------